Bookmarks E2E tests.

+++ src/bookmark/bookmark.service.ts
@@ -1,14 +1,20 @@
-import { Injectable } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+} from '@nestjs/common';
 import { PrismaService } from '../prisma/prisma.service';
-import { CreateBookmarkDto } from './dto';
+import {
+  CreateBookmarkDto,
+  EditBookmarkDto,
+} from './dto';
 
 @Injectable()
 export class BookmarkService {
   constructor(private prisma: PrismaService) {}
-  getBookmarks(id: number) {
+  getBookmarks(userId: number) {
     return this.prisma.bookmark.findMany({
       where: {
-        userId: id,
+        userId,
       },
     });
   }
@@ -22,15 +28,34 @@ export class BookmarkService {
     });
   }
 
-  editBookmarkById(
+  createBookmark(
     userId: number,
-    id: number,
     dto: CreateBookmarkDto,
   ) {
+    return this.prisma.bookmark.create({
+      data: { userId, ...dto },
+    });
+  }
+
+  async editBookmarkById(
+    userId: number,
+    id: number,
+    dto: EditBookmarkDto,
+  ) {
+    const bookmark =
+      await this.prisma.bookmark.findUnique({
+        where: {
+          id,
+        },
+      });
+    if (!bookmark || bookmark.userId !== userId)
+      throw new ForbiddenException(
+        'Access to resource denied',
+      );
+
     return this.prisma.bookmark.update({
       where: {
         id,
-        userId,
       },
       data: {
         ...dto,
@@ -38,18 +63,24 @@ export class BookmarkService {
     });
   }
 
-  deleteBookmarkById(id: number) {
-    return this.prisma.bookmark.delete({
-      where: { id },
-    });
-  }
-
-  createBookmark(
+  async deleteBookmarkById(
     userId: number,
-    dto: CreateBookmarkDto,
+    id: number,
   ) {
-    return this.prisma.bookmark.create({
-      data: { userId, ...dto },
+    const bookmark =
+      await this.prisma.bookmark.findUnique({
+        where: {
+          id,
+        },
+      });
+
+    if (!bookmark || bookmark.userId !== userId)
+      throw new ForbiddenException(
+        'Access to resource denied!',
+      );
+
+    await this.prisma.bookmark.delete({
+      where: { id, userId },
     });
   }
 }
